requirements.txt
# Disaster Relief Resource Optimizer
# Dependencies for the complete application

# Core data science libraries
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2

# Visualization libraries
matplotlib==3.8.1
seaborn==0.13.0
plotly==5.17.0

# Web framework
streamlit==1.28.2

# Authentication and database
supabase==2.0.2
streamlit-authenticator==0.2.3

# Additional utilities
requests==2.31.0
python-dotenv==1.0.0

# Development and testing (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0